Several changes in upgrade

See merge request automation/pg-tests!938

+++ tests_install/test_upgrade.py
@@ -6,7 +6,8 @@ import allure
 from allure_commons.types import LabelType
 from helpers.pginstall import PgInstall, PGPRO_ARCHIVE_ENTERPRISE,\
     PGPRO_ARCHIVE_STANDARD, ALT_BASED, DEBIAN_BASED
-from helpers.constants import FIRST_RELEASE
+from helpers.constants import FIRST_RELEASE, UPGRADE_ROUTES,\
+    DUMP_RESTORE_ROUTES
 from helpers.utils import diff_dbs, download_dump, urlopen, get_distro, \
     compare_versions, extend_ver
 import time
@@ -21,369 +22,6 @@ import re
 
 system = platform.system()
 tempdir = tempfile.gettempdir()
-
-UPGRADE_ROUTES = {
-
-    'postgrespro-std-9.6': {
-        'from': [
-            {
-                'name': 'postgresql', 'edition': '', 'version': '9.6',
-                'initdb-params': '--locale=C'
-            }
-        ]
-    },
-
-    'postgrespro-std-10': {
-        'from': [
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '9.6'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '10',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '9.6',
-                'initdb-params': '--locale=C'
-            },
-        ]
-    },
-
-    'postgrespro-std-11': {
-        'from': [
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '10'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '9.6'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '11',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '10',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '9.6',
-                'initdb-params': '--locale=C'
-            },
-        ]
-    },
-    'postgrespro-std-12': {
-        'from': [
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '11'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '10'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '9.6'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '11',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '10',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '9.6',
-                'initdb-params': '--locale=C'
-            },
-        ]
-    },
-    'postgrespro-std-cert-11': {
-        'from': [
-            {
-                'name': 'postgrespro', 'edition': 'std-cert', 'version': '10'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '10'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '9.6'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '11',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '10',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '9.6',
-                'initdb-params': '--locale=C'
-            },
-        ]
-    },
-
-    'postgrespro-ent-10': {
-        'from': [
-            {
-                'name': 'postgrespro', 'edition': 'ent', 'version': '9.6'
-            }
-        ]
-    },
-
-    'postgrespro-ent-11': {
-        'from': [
-            {
-                'name': 'postgrespro', 'edition': 'ent', 'version': '9.6'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'ent', 'version': '10'
-            }
-        ]
-    },
-
-    'postgrespro-ent-12': {
-        'from': [
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '12'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '11'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '10'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '9.6'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'ent', 'version': '11'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'ent', 'version': '10'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'ent', 'version': '9.6'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '12',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '11',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '10',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '9.6',
-                'initdb-params': '--locale=C'
-            },
-        ]
-    }
-
-}
-
-DUMP_RESTORE_ROUTES = {
-
-    'postgrespro-std-9.6': {
-        'from': [
-            {
-                'name': 'postgresql', 'edition': '', 'version': '9.6',
-                'initdb-params': '--locale=C'
-            }
-        ]
-    },
-
-    'postgrespro-std-10': {
-        'from': [
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '9.6'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '10',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '9.6',
-                'initdb-params': '--locale=C'
-            },
-        ]
-    },
-
-    'postgrespro-std-11': {
-        'from': [
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '10'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '9.6'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '11',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '10',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '9.6',
-                'initdb-params': '--locale=C'
-            },
-        ]
-    },
-    'postgrespro-std-12': {
-        'from': [
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '11'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '10'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '9.6'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '11',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '10',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '9.6',
-                'initdb-params': '--locale=C'
-            },
-        ]
-    },
-    'postgrespro-std-cert-11': {
-        'from': [
-            {
-                'name': 'postgrespro', 'edition': 'std-cert', 'version': '10'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '10'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '9.6'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '11',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '10',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '9.6',
-                'initdb-params': '--locale=C'
-            },
-        ]
-    },
-
-    'postgrespro-ent-10': {
-        'from': [
-            {
-                'name': 'postgrespro', 'edition': 'ent', 'version': '9.6'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '9.6'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '10'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '10',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '9.6',
-                'initdb-params': '--locale=C'
-            }
-        ]
-    },
-
-    'postgrespro-ent-11': {
-        'from': [
-            {
-                'name': 'postgrespro', 'edition': 'ent', 'version': '9.6'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'ent', 'version': '10'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '9.6'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '10'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '11'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '10',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '11',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '9.6',
-                'initdb-params': '--locale=C'
-            }
-        ]
-    },
-
-    'postgrespro-ent-12': {
-        'from': [
-            {
-                'name': 'postgrespro', 'edition': 'ent', 'version': '9.6'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'ent', 'version': '10'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'ent', 'version': '11'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '9.6'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '10'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '11'
-            },
-            {
-                'name': 'postgrespro', 'edition': 'std', 'version': '12'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '10',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '11',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '12',
-                'initdb-params': '--locale=C'
-            },
-            {
-                'name': 'postgresql', 'edition': '', 'version': '9.6',
-                'initdb-params': '--locale=C'
-            }
-        ]
-    }
-
-}
-
-
 upgrade_dir = os.path.join(tempfile.gettempdir(), 'upgrade')
 amcheck_sql = """
 create extension if not exists amcheck;
@@ -431,13 +69,16 @@ BEGIN
     LOOP
         EXECUTE alter_command;
     END LOOP;
+    DROP VIEW IF EXISTS my_locks;
 END;
 $$;
 """
-drop_oids_sql = """
+prepare_for_12_plus_sql = """
 DO $$
 DECLARE
     table_name TEXT;
+    attname TEXT;
+    typname TEXT;
 BEGIN
     FOR table_name IN
         SELECT '"' || n.nspname || '"."' || c.relname || '"' AS tab
@@ -447,6 +88,45 @@ BEGIN
             NOT IN ('pg_catalog') order by c.oid
     LOOP
         EXECUTE 'ALTER TABLE ' || table_name || ' SET WITHOUT OIDS';
+    END LOOP;
+    FOR table_name, attname, typname IN
+        SELECT
+            '"' || n.nspname || '"."' || c.relname || '"' AS tab,
+            pa.attname as col,
+            pt.typname as typname
+        FROM
+            pg_catalog.pg_class c
+            JOIN pg_catalog.pg_namespace n ON c.relnamespace = n.oid
+            JOIN pg_catalog.pg_attribute pa ON pa.attrelid=c.oid
+            JOIN pg_catalog.pg_type pt ON pa.atttypid=pt.oid
+        WHERE
+            n.nspname NOT IN ('pg_catalog')
+            AND pt.typname IN ('abstime','reltime','tinterval','smgr',
+                '_abstime','_reltime','_tinterval','_smgr')
+        ORDER BY c.oid
+    LOOP
+        EXECUTE 'ALTER TABLE ' || table_name || ' ALTER COLUMN ' || attname ||
+         ' TYPE VARCHAR';
+    END LOOP;
+END;
+$$;
+"""
+prepare_for_13_plus_sql = """
+DO $$
+DECLARE
+    con_name TEXT;
+BEGIN
+    FOR con_name IN
+        SELECT
+            '"' || n.nspname || '"."' || pc.conname || '"' AS cname
+        FROM pg_catalog.pg_conversion pc
+            JOIN pg_catalog.pg_namespace n ON pc.connamespace = n.oid
+        WHERE
+            n.nspname NOT IN ('pg_catalog')
+            AND 'SQL_ASCII' IN (pg_encoding_to_char(pc.conforencoding),
+                                pg_encoding_to_char(pc.contoencoding))
+    LOOP
+        EXECUTE 'DROP CONVERSION ' || con_name || ';';
     END LOOP;
 END;
 $$;
@@ -516,6 +196,7 @@ def install_server(product, edition, version, milestone, branch, windows,
         pg.install_postgres_win()
         pg.client_path_needed = True
         pg.server_path_needed = True
+        pg.install_default_config()
         pg.load_shared_libraries()
     return pg
 
@@ -525,11 +206,18 @@ def generate_db(pg, pgnew, custom_dump=None):
     dump_file_name = download_dump(pg.product, pg.edition, pg.version,
                                    tempdir, custom_dump)
     with open(os.path.join(tempdir, 'load-%s.log' % key), 'wb') as out:
-        pg.exec_psql_file(dump_file_name, '-q',
+        pg.exec_psql_file(dump_file_name, '-q -v ON_ERROR_STOP=1',
                           stdout=out)
-    if pgnew.version not in ["9.6", "10", "11"] and \
-            pg.version in ["9.6", "10", "11"]:
-        pg.do_in_all_dbs(drop_oids_sql)
+    if compare_versions(pg.version, '12') < 0 and \
+            compare_versions(pgnew.version, '12') >= 0:
+        pg.do_in_all_dbs(prepare_for_12_plus_sql)
+    if compare_versions(pg.version, '13') < 0 and \
+            compare_versions(pgnew.version, '13') >= 0:
+        pg.do_in_all_dbs(prepare_for_13_plus_sql)
+    # wait for 12.5
+    if pg.version == '12':
+        pg.exec_psql('DROP TABLE IF EXISTS test_like_5c CASCADE',
+                     '-d regression')
     if pgnew.edition in ['ent', 'ent-cert'] and \
             pg.edition not in ['ent', 'ent-cert']:
         pg.do_in_all_dbs(remove_xid_type_columns_sql)
@@ -622,6 +310,7 @@ def init_cluster(pg, force_remove=True, initdb_params='',
     else:
         stop(pg, stopped)
         pg.init_cluster(force_remove, '-k ' + initdb_params)
+        pg.install_default_config()
         start(pg)
         if load_libs:
             pg.load_shared_libraries(restart_service=False)
@@ -773,7 +462,8 @@ class TestUpgrade():
                 ])
             # PGPRO-2459
             if pgold.os_name in DEBIAN_BASED and \
-                    old_name == "postgrespro" and old_version == "9.6":
+                    old_name == "postgrespro" and old_version == "9.6" and \
+                    old_edition != '1c':
                 subprocess.check_call("apt-get purge -y postgrespro-common "
                                       "postgrespro-client-common", shell=True)
 
@@ -860,7 +550,7 @@ class TestUpgrade():
                 with open(os.path.join(tempdir, 'load-dr-%s.log' % old_key),
                           'wb') as out:
                     pg.exec_psql_file(file_name, '-q', stdout=out)
-                dump_and_diff_dbs(old_key, pg, 'upgrade')
+                dump_and_diff_dbs(old_key, pg, 'dump-restore')
                 stop(pg)
                 pgold.remove_full(True, do_not_remove=[
                     r"^libc.*", r".*icu.*", r".*zstd.*", r"^llvm.*"
